fix: use u32 for SQLite count column (usize has no FromSql impl)
Namespace count query returned integer from SQLite but tried to
deserialize as usize — rusqlite doesn't implement FromSql for
usize. Use u32 which is safe for any realistic count and supported
by rusqlite directly.

+++ crates/uteke-core/src/memory/tags.rs
@@ -372,7 +372,7 @@ impl super::Store {
 
         let rows = stmt
             .query_map([], |row: &rusqlite::Row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, usize>(1)?))
+                Ok((row.get::<_, String>(0)?, row.get::<_, u32>(1)? as usize))
             })
             .map_err(|e| Error::db("database operation", e))?;
